Fixes index query for references with no hits (htslib d7541514).
This didn't actually affect io_lib as the seek function operates
differently and returns a different error code, but the additional
check does not harm and brings the implementations into alignment
again.

Also fixed cram_fd->record_counter to be 64-bit; an omission in the
recent commits.

+++ io_lib/cram_structs.h
@@ -746,7 +746,7 @@ typedef struct {
     SAM_hdr       *header;
 
     char          *prefix;
-    int            record_counter;
+    int64_t        record_counter;
     int            err;
 
     // Most recent compression header decoded
